refactoring: core namespace with abstracts
Первая часть преобразований -- вынесены абстрактные части.
Осталось завести по реализации для каждого языка и заняться фабриками.

Я НЕ ДЕЛАЛА ОТДЕЛЬНЫЕ ФАЙЛЫ ЧТОБЫ СОХРАНИТЬ ЧИТАЕМОСТЬ
