+++ task_2/classunit.cpp
@@ -2,17 +2,7 @@
 
 #include <sstream>
 
-const std::vector<std::string> ClassUnit::ACCESS_MODIFIERS = { "public",
-                                                               "protected",
-                                                               "private" };
-
-ClassUnit::ClassUnit(const std::string& name)
-    : m_name(name)
-{
-    m_fields.resize(ACCESS_MODIFIERS.size());
-}
-
-void ClassUnit::add(const std::shared_ptr<Unit> &unit, Unit::Flags flags)
+void CppClassUnit::add(const std::shared_ptr<Unit> &unit, Unit::Flags flags)
 {
     auto accessModifier = PRIVATE;
     if(flags < ACCESS_MODIFIERS.size()) {
@@ -21,7 +11,7 @@ void ClassUnit::add(const std::shared_ptr<Unit> &unit, Unit::Flags flags)
     m_fields[accessModifier].push_back(unit);
 }
 
-std::string ClassUnit::compile(unsigned int level) const
+std::string CppClassUnit::compile(unsigned int level) const
 {
     std::stringstream result;
     result << core::generateShift(level)
