+++ task_2/methodunit.h
@@ -4,25 +4,12 @@
 #include "unit.h"
 #include <vector>
 
-class MethodUnit : public core::Unit {
+class CppMethodUnit : public core::MethodUnit {
 public:
-    enum Modifier {
-        STATIC = 1,
-        CONST = 1 << 1,
-        VIRTUAL = 1 << 2
-    };
-public:
-    MethodUnit(const std::string& name, const std::string& returnType, Flags flags);
+    using MethodUnit::MethodUnit;
 
     void add(const std::shared_ptr< Unit >& unit, Flags /* flags */ = 0);
 
     std::string compile( unsigned int level = 0 ) const;
-
-private:
-    std::string m_name;
-    std::string m_returnType;
-
-    Flags m_flags;
-    std::vector<std::shared_ptr<Unit>> m_body;
 };
 #endif // METHODUNIT_H
