+++ task_2/printoperatorunit.cpp
@@ -2,11 +2,7 @@
 
 #include <sstream>
 
-PrintOperatorUnit::PrintOperatorUnit(const std::string &text)
-    : m_text(text)
-{ }
-
-std::string PrintOperatorUnit::compile(unsigned int level) const
+std::string CppPrintOperatorUnit::compile(unsigned int level) const
 {
     std::stringstream ss;
     ss << core::generateShift(level)
